requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
Babel==2.15.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
courlan==1.1.0
dateparser==1.2.0
distro==1.9.0
Django==4.2
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
Faker==24.14.1
h11==0.14.0
htmldate==1.8.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jusText==3.0.0
lxml==5.1.1
openai==1.25.1
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.28
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tld==0.13
toposort==1.10
tqdm==4.66.4
trafilatura==1.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
